add admin permission

+++ admin/categories/index.php
@@ -1,4 +1,7 @@
-<?php session_start() ?>
+<?php 
+    session_start();
+    require '../check_admin.php'
+?>
 <!DOCTYPE html>
 <html lang="en">
 <head>
@@ -19,6 +22,8 @@
     require '../connect.php';
     $sql = "select * from categories order by id desc";
     $result = mysqli_query($connection, $sql);
+    mysqli_close($connection);
+
 ?>
 <div class="admin-main">
     <?php include '../topbar.php' ?>
@@ -54,7 +59,7 @@
                             <a href="./form_update.php?id=<?php echo $each['id'] ?>">Sửa</a>
                         </td>
                         <td class="table-form-item">
-                            <a href="./delete.php?id=<?php echo $each['id'] ?>">Xóa</a>
+                            <a onclick="return confirm('Bạn chắc chắn muốn xóa thể loại này chứ')" href="./delete.php?id=<?php echo $each['id'] ?>">Xóa</a>
                         </td>
                     </tr>
                     <?php } ?>
