+++ admin/manufacturers/index.php
@@ -1,4 +1,7 @@
-<?php session_start() ?>
+<?php 
+    session_start();
+    require '../check_super_admin.php';
+?>
 <!DOCTYPE html>
 <html lang="en">
 <head>
@@ -19,6 +22,8 @@
     require '../connect.php';
     $sql = "select * from manufacturers order by id asc";
     $result = mysqli_query($connection, $sql);
+    mysqli_close($connection);
+
 ?>
 <div class="admin-main">
     <?php include '../topbar.php' ?>
@@ -50,7 +55,7 @@
                             <a href="./form_update.php?id=<?php echo $each['id'] ?>">Sửa</a>
                         </td>
                         <td class="table-form-item">
-                            <a href="./delete.php?id=<?php echo $each['id'] ?>">Xóa</a>
+                            <a onclick="return confirm('Bạn chắc chắn muốn xóa nhà sản xuất này chứ')" href="./delete.php?id=<?php echo $each['id'] ?>">Xóa</a>
                         </td>
                     </tr>
                     <?php } ?>
