+++ admin/staffs/delete.php
@@ -1,4 +1,5 @@
 <?php 
+    require '../check_super_admin.php';
 
 if(empty($_GET['id'])){
     $_SESSION['error'] = 'Phải truyền mã để xóa';
@@ -8,7 +9,11 @@ if(empty($_GET['id'])){
 
 $id = $_GET['id'];
 
-
+if($id == $_SESSION['id']){
+    $_SESSION['error'] = 'Không thể xóa bản thân';
+    header('location:index.php');
+    exit;
+}
 
 require '../connect.php';
 $sql = "delete from staffs where id = '$id'";
